Moving IO functions that were defined in NanoOsLibC to NanoOsIo.
